* Add a scratch implementation of the new inventory fields to the mssql database adapter * I don't use mssql so this may not work, corrections (in the form of patches) are welcome. * Unlike mysql, mssql requires manual updating of existing tables here (which should mean just adding the new fields manually)

+++ OpenSim/Data/MSSQL/Resources/CreateItemsTable.sql
@@ -19,10 +19,10 @@ CREATE TABLE [inventoryitems] (
   [inventoryBasePermissions] [int] NOT NULL default 0,
   [inventoryEveryOnePermissions] [int] NOT NULL default 0,
   [salePrice] [int] default NULL,
-  [saleType] [smallint] default NULL,
+  [saleType] [tinyint] default NULL,
   [creationDate] [int] default NULL,
   [groupID] [varchar](36) default NULL,
-  [groupOwned] [smallint] default NULL,
+  [groupOwned] [tinyint] default NULL,
   [flags] [int] unsigned default NULL,
   PRIMARY KEY CLUSTERED
 (
